traitement_mcf/requirements.txt
dateparser==1.2.0
pandas==2.2.2
python-dotenv==1.0.1
regex==2024.7.24
Requests==2.32.3
SQLAlchemy==2.0.32
pyarrow==17.0.0
urllib3==2.2.2
psycopg2-binary==2.9.9
regex